+++ CHANGELOG.md
@@ -11,5 +11,6 @@ All notable changes to this project will be documented in this file.
 ## [0.0.2] - 2019-02-18
 ### Feature
 - Order data model with update behaviors
+- Exchange place and cancel order methods for execution